OneDrive/Desktop/FJD_Project/Backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow-cpu
numpy
pandas
requests
Pillow
firebase-admin
python-Levenshtein
dnspython
email-validator